Added the provider pattern for the forecast data, along with dependency injection
