+++ DataService/Card App.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27703.2047
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28803.202
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Card App", "Card App.csproj", "{D8A991D1-864D-4392-B545-2391B4A1A282}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CardApp", "CardApp.csproj", "{D8A991D1-864D-4392-B545-2391B4A1A282}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
