+++ DataService/Controllers/SampleDataController.cs
@@ -1,44 +1,24 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using CardApp.Models;
+using CardApp.Providers.Interfaces;
 using Microsoft.AspNetCore.Mvc;
+using System.Collections.Generic;
 
-namespace Card_App.Controllers
+namespace CardApp.Controllers
 {
     [Route("api/[controller]")]
     public class SampleDataController : Controller
     {
-        private static string[] Summaries = new[]
+        private readonly IWeatherForecastProvider weatherForecastProvider;
+
+        public SampleDataController(IWeatherForecastProvider weatherForecastProvider)
         {
-            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-        };
+            this.weatherForecastProvider = weatherForecastProvider;
+        }        
 
         [HttpGet("[action]")]
         public IEnumerable<WeatherForecast> WeatherForecasts(int startDateIndex)
         {
-            var rng = new Random();
-            return Enumerable.Range(1, 5).Select(index => new WeatherForecast
-            {
-                DateFormatted = DateTime.Now.AddDays(index + startDateIndex).ToString("d"),
-                TemperatureC = rng.Next(-20, 55),
-                Summary = Summaries[rng.Next(Summaries.Length)]
-            });
-        }
-
-        public class WeatherForecast
-        {
-            public string DateFormatted { get; set; }
-            public int TemperatureC { get; set; }
-            public string Summary { get; set; }
-
-            public int TemperatureF
-            {
-                get
-                {
-                    return 32 + (int)(TemperatureC / 0.5556);
-                }
-            }
+            return this.weatherForecastProvider.WeatherForecasts(startDateIndex);
         }
     }
 }
