+++ DataService/Models/WeatherForecast.cs
@@ -1,0 +1,17 @@
+﻿namespace CardApp.Models
+{
+    public class WeatherForecast
+    {
+        public string DateFormatted { get; set; }
+        public int TemperatureC { get; set; }
+        public string Summary { get; set; }
+
+        public int TemperatureF
+        {
+            get
+            {
+                return 32 + (int)(TemperatureC / 0.5556);
+            }
+        }
+    }
+}
